Manga file/requirements.txt
aiogram>=3.0.0
requests>=2.28.0
tenacity>=8.0.0
img2pdf>=0.4.0
Pillow>=9.0.0
telegraph>=2.0.0
aiosqlite>=0.19.0
